Support IServiceCollection provided by Functions (#103)
* Just hacking to get it going (no tests, hardcoding)

* Update GitVersion.yml

* Extract configuration extension

* Merge ServerlessEndpoint into FunctionEndpoint
