+++ GitVersion.yml
@@ -2,7 +2,7 @@ assembly-versioning-scheme: Major
 branches:
   master:
     mode: ContinuousDeployment
-    tag: alpha
+    tag: experimental
     increment: Minor
     tracks-release-branches: true
   release:
